Train code split, Part 4: moving some of the model.py code into dataset.py (#40)
Moving data related code to dataset.py, adding docstrings and type hints where missing and doing minor refactorings (major ones coming up next).

Moving data_loaders.py to the clavaddpm folder and adding missing docstrings and type hints in that file.

Also adding missing docstrings and type hints in model.py.
