+++ src/midst_toolkit/models/clavaddpm/typing.py
@@ -1,4 +1,8 @@
+from collections.abc import Callable
 from typing import Any
+
+import numpy as np
+from torch import nn
 
 
 # TODO: Temporary, will switch to classes later
@@ -6,3 +10,5 @@ Configs = dict[str, Any]
 Tables = dict[str, dict[str, Any]]
 RelationOrder = list[tuple[str, str]]
 GroupLengthsProbDicts = dict[tuple[str, str], dict[int, dict[int, float]]]
+ArrayDict = dict[str, np.ndarray]
+ModuleType = str | Callable[..., nn.Module]
