+++ tests/integration/models/clavaddpm/test_model.py
@@ -10,8 +10,8 @@ import torch
 from torch.nn import functional
 
 from midst_toolkit.common.random import set_all_random_seeds, unset_all_random_seeds
-from midst_toolkit.core.data_loaders import load_multi_table
 from midst_toolkit.models.clavaddpm.clustering import clava_clustering
+from midst_toolkit.models.clavaddpm.data_loaders import load_multi_table
 from midst_toolkit.models.clavaddpm.model import Classifier
 from midst_toolkit.models.clavaddpm.train import clava_training
 
@@ -46,7 +46,7 @@ CLASSIFIER_CONFIG = {
 
 @pytest.mark.integration_test()
 def test_load_single_table():
-    tables, relation_order, dataset_meta = load_multi_table("tests/integration/assets/single_table/")
+    tables, relation_order, dataset_meta = load_multi_table(Path("tests/integration/assets/single_table/"))
 
     assert list(tables.keys()) == ["trans"]
 
@@ -121,7 +121,7 @@ def test_load_single_table():
 
 @pytest.mark.integration_test()
 def test_load_multi_table():
-    tables, relation_order, dataset_meta = load_multi_table("tests/integration/assets/multi_table/")
+    tables, relation_order, dataset_meta = load_multi_table(Path("tests/integration/assets/multi_table/"))
 
     assert list(tables.keys()) == ["account", "trans"]
 
@@ -251,7 +251,7 @@ def test_train_single_table(tmp_path: Path):
     set_all_random_seeds(seed=133742, use_deterministic_torch_algos=True, disable_torch_benchmarking=True)
 
     # Act
-    tables, relation_order, _ = load_multi_table("tests/integration/assets/single_table/")
+    tables, relation_order, _ = load_multi_table(Path("tests/integration/assets/single_table/"))
     tables, models = clava_training(
         tables, relation_order, tmp_path, DIFFUSION_CONFIG, CLASSIFIER_CONFIG, device="cpu"
     )
@@ -309,7 +309,7 @@ def test_train_multi_table(tmp_path: Path):
     set_all_random_seeds(seed=133742, use_deterministic_torch_algos=True, disable_torch_benchmarking=True)
 
     # Act
-    tables, relation_order, _ = load_multi_table("tests/integration/assets/multi_table/")
+    tables, relation_order, _ = load_multi_table(Path("tests/integration/assets/multi_table/"))
     tables, all_group_lengths_prob_dicts = clava_clustering(tables, relation_order, tmp_path, CLUSTERING_CONFIG)
     models = clava_training(tables, relation_order, tmp_path, DIFFUSION_CONFIG, CLASSIFIER_CONFIG, device="cpu")
 
@@ -395,7 +395,7 @@ def test_clustering_reload(tmp_path: Path):
     set_all_random_seeds(seed=133742, use_deterministic_torch_algos=True, disable_torch_benchmarking=True)
 
     # Act
-    tables, relation_order, dataset_meta = load_multi_table("tests/integration/assets/multi_table/")
+    tables, relation_order, dataset_meta = load_multi_table(Path("tests/integration/assets/multi_table/"))
     tables, all_group_lengths_prob_dicts = clava_clustering(tables, relation_order, tmp_path, CLUSTERING_CONFIG)
 
     # Assert
